Fix customer front-end routing and trending data

+++ Presentation/Categorie/supprimerCategorie.php
@@ -1,13 +1,17 @@
-<?php 
-  session_start();
-  include_once('../../Metier/categorie.php');
-  if(!isset($_SESSION['login'])){
-    header("Location: http://localhost/Mini/");
-  }
+<?php
+require_once __DIR__ . '/../../config/app.php';
+require_once __DIR__ . '/../../Metier/categorie.php';
 
+session_start();
 
-        Categorie::delete($_GET['id']);
-  
+if (!isset($_SESSION['user'])) {
+    header('Location: ' . url_for('index.php'));
+    exit();
+}
 
-    header("Location: http://localhost/Mini/Presentation/Categorie/afficherCategories.php");
-?>
+if (isset($_GET['id'])) {
+    Categorie::delete($_GET['id']);
+}
+
+header('Location: ' . url_for('Presentation/Categorie/afficherCategories.php'));
+exit();
